add HTTP example

+++ src/ATOM_DTU_NB.h
@@ -7,6 +7,6 @@ class ATOM_DTU_NB
         HardwareSerial *_serial;
     public:
         void Init(HardwareSerial *serial = &Serial2, uint8_t RX = 19, uint8_t TX = 22);
-        String waitMsg(int time);
+        String waitMsg(unsigned long time);
         void sendMsg(String command);
 };
